MDL-60910 question: make drag item backgroup semi-transparent
This commit changes the background color of draggable items in the
drag and drop image question type to semi transparent. This enables
the use of transparent pictures in the drag and drop question type.

Before this change it was not possible to e.g. place countries on a
map because the background map was not visible anymore, since the
draggable image items will have a colored background themself.

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>