MDL-72708 template: backslash should be quoted to produce valid json
The rendering of the aria-labels in
mod/forum/templates/forum_discussion_post.mustache used json-encoded
variables as arguments working on user provided input. If this contained
a backslash, it produced invalid JSON syntax.

This change adds the {{#quote}} helper to the forum discussion templates
and for rendingering the aria-labels and also extends the quote helper
to properly escape backslashes.