MDL-81669 libraries: Upgrade phpxmlrpc to 4.11.0
I also removed the version specific parts of readme_moodle.txt as those
infos are already available in the thirdpartylibs.xml. Also the
mentioned local changes are no longer relevant, so they can go too.

And I added the missing composer/PSR4 and moved the lib to the correct
location according to MDL-82829 and MDL-82747.

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>