MDL-76512 question: Fix moving categories
The polyfill core-js-bundle had erroneously added `forEach` on all DOM
collection prototypes. The element in question is a `HTMLCollection`
which is not an array but an array-like object. We use `getList()`
to convert the array-like object to a real array and then use `forEach`.
See https://github.com/zloirock/core-js/issues/988 for more details.

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>