MDL-77593 communication: improve performance adding or removing members
Before this change, the communication room gets all the users when
enrolled or from the course when a new communication instance is added.
It happens using an ad-hoc task but there is no efficiency to handle the
number of users should be added in one go. This might have scaling
issues in terms of courses with massive number of enrolment or users.

This commit implements a logic to chunk the data and queue them so these
can happen in smaller chunks without keeping the servers busy for a long
time.

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>