MDL-72900 mod_choice: fix styling of long options
If the text in the choice option is very long, the radio button (or the
checkbox) takes one row while the option itself goes to another, like
there is a line break between them. We fix this using CSS Grids putting
the input and the lable into their own columns.

I also removed the `display: inline-block;` from `.path-mod-choice
.horizontal .choices .option` since the display mode should only affect
the display of the results (see MDL-59843 and MDL-37946 for reference).

Signed-off-by: Daniel Ziegenberg <daniel@ziegenberg.at>